Moved webui dev app to src, using it as 'default' app for the Docker image
