+++ .vscode/settings.json
@@ -3,5 +3,6 @@
     "typescript.tsdk": "webui/node_modules/typescript/lib",
     "eslint.options": {
         "configFile": "webui/configs/eslintrc.json"
-    }
+    },
+    "eslint.packageManager": "yarn"
 }
