+++ cli/package.json
@@ -52,7 +52,7 @@
         "clean": "rimraf lib",
         "build": "tsc -p ./tsconfig.json && yarn run lint && cp src/ovsx lib/ovsx",
         "watch": "tsc -w -p ./tsconfig.json",
-        "lint": "eslint -c ../webui/configs/eslintrc.json --ext .ts src",
+        "lint": "eslint -c ./configs/eslintrc.json --ext .ts src",
         "prepare": "yarn run clean && yarn run build",
         "publish:next": "yarn publish --new-version \"$(semver $npm_package_version -i minor)-next.$(git rev-parse --short HEAD)\" --tag next --no-git-tag-version",
         "publish:latest": "yarn publish --tag latest"
