+++ webui/package.json
@@ -46,11 +46,16 @@
         "react-infinite-scroller": "^1.2.4",
         "react-router-dom": "^5.2.0"
     },
+    "peerDependencies": {
+        "express": "^4.17.1"
+    },
     "devDependencies": {
         "@types/chai": "^4.2.11",
         "@types/dompurify": "^2.0.2",
+        "@types/express": "^4.17.7",
         "@types/markdown-it": "^10.0.0",
         "@types/mocha": "^8.0.0",
+        "@types/node": "12.12.53",
         "@types/punycode": "^2.1.0",
         "@types/react": "^16.9.34",
         "@types/react-dom": "^16.9.6",
@@ -79,9 +84,9 @@
         "lint": "eslint -c ./configs/eslintrc.json --ext .ts,.tsx src",
         "watch": "tsc -w -p ./tsconfig.json",
         "prepare": "yarn run clean && yarn run build",
-        "build:dev": "webpack --config ./configs/webpack.config.js",
-        "watch:dev": "webpack --watch --config ./configs/webpack.config.js",
-        "start:dev": "node dev/server/main",
+        "build:default": "webpack --config ./configs/webpack.config.js --mode production",
+        "watch:default": "webpack --config ./configs/webpack.config.js --mode development --watch",
+        "start:default": "node lib/default/server",
         "publish:next": "yarn publish --new-version \"$(semver $npm_package_version -i minor)-next.$(git rev-parse --short HEAD)\" --tag next --no-git-tag-version",
         "publish:latest": "yarn publish --tag latest"
     }
