+++ webui/src/components/copy-to-clipboard.tsx
