+++ webui/src/components/text-divider.tsx
