+++ webui/src/pages/extension-detail/extension-detail-reviews.tsx
@@ -12,10 +12,10 @@ import * as React from 'react';
 import { Theme, createStyles, WithStyles, withStyles, Box, Typography, Divider, Link } from '@material-ui/core';
 import { toLocalTime } from '../../utils';
 import { ExtensionReview, UserData, Extension, ExtensionReviewList, isEqualUser, isError } from '../../extension-registry-types';
-import { TextDivider } from '../../custom-mui-components/text-divider';
-import { DelayedLoadIndicator } from '../../custom-mui-components/delayed-load-indicator';
-import { ButtonWithProgress } from '../../custom-mui-components/button-with-progress';
-import { Timestamp } from '../../custom-mui-components/timestamp';
+import { TextDivider } from '../../components/text-divider';
+import { DelayedLoadIndicator } from '../../components/delayed-load-indicator';
+import { ButtonWithProgress } from '../../components/button-with-progress';
+import { Timestamp } from '../../components/timestamp';
 import { ExtensionRegistryService } from '../../extension-registry-service';
 import { ExportRatingStars } from './extension-rating-stars';
 import { ExtensionReviewDialog } from './extension-review-dialog';
