+++ webui/src/pages/extension-detail/extension-detail.tsx
@@ -16,11 +16,11 @@ import VerifiedUserIcon from '@material-ui/icons/VerifiedUser';
 import PublicIcon from '@material-ui/icons/Public';
 import WarningIcon from '@material-ui/icons/Warning';
 import { createRoute } from '../../utils';
-import { DelayedLoadIndicator } from '../../custom-mui-components/delayed-load-indicator';
-import { HoverPopover } from '../../custom-mui-components/hover-popover';
+import { DelayedLoadIndicator } from '../../components/delayed-load-indicator';
+import { HoverPopover } from '../../components/hover-popover';
 import { ExtensionRegistryService } from '../../extension-registry-service';
 import { Extension, UserData, isError } from '../../extension-registry-types';
-import { TextDivider } from '../../custom-mui-components/text-divider';
+import { TextDivider } from '../../components/text-divider';
 import { PageSettings } from '../../page-settings';
 import { ExtensionDetailOverview } from './extension-detail-overview';
 import { ExtensionDetailReviews } from './extension-detail-reviews';
