+++ webui/src/pages/extension-detail/extension-review-dialog.tsx
@@ -11,7 +11,7 @@
 import * as React from 'react';
 import { Button, Dialog, DialogTitle, DialogContent, DialogContentText, TextField, DialogActions, Theme } from '@material-ui/core';
 import { withStyles, createStyles, WithStyles } from '@material-ui/styles';
-import { ButtonWithProgress } from '../../custom-mui-components/button-with-progress';
+import { ButtonWithProgress } from '../../components/button-with-progress';
 import { ExtensionRegistryService } from '../../extension-registry-service';
 import { UserData, Extension, isError } from '../../extension-registry-types';
 import { ExtensionRatingStarSetter } from './extension-rating-star-setter';
