+++ webui/src/pages/extension-list/extension-list.tsx
@@ -15,7 +15,7 @@ import { ExtensionListItem } from './extension-list-item';
 import { isError, SearchEntry } from '../../extension-registry-types';
 import { ExtensionRegistryService, ExtensionFilter } from '../../extension-registry-service';
 import { debounce } from '../../utils';
-import { DelayedLoadIndicator } from '../../custom-mui-components/delayed-load-indicator';
+import { DelayedLoadIndicator } from '../../components/delayed-load-indicator';
 import { PageSettings } from '../../page-settings';
 import { ErrorResponse } from '../../server-request';
