+++ webui/src/pages/user/user-settings-tokens.tsx
@@ -10,8 +10,8 @@
 
 import * as React from 'react';
 import { Theme, createStyles, WithStyles, withStyles, Typography, Box, Paper, Button } from '@material-ui/core';
-import { DelayedLoadIndicator } from '../../custom-mui-components/delayed-load-indicator';
-import { Timestamp } from '../../custom-mui-components/timestamp';
+import { DelayedLoadIndicator } from '../../components/delayed-load-indicator';
+import { Timestamp } from '../../components/timestamp';
 import { UserData, PersonalAccessToken } from '../../extension-registry-types';
 import { ExtensionRegistryService } from '../../extension-registry-service';
 import { GenerateTokenDialog } from './generate-token-dialog';
