+++ webui/src/pages/user/user-settings.tsx
@@ -12,7 +12,7 @@ import * as React from 'react';
 import { createStyles, Theme, WithStyles, withStyles, Grid, Container, Box, Typography } from '@material-ui/core';
 import { RouteComponentProps, Route } from 'react-router-dom';
 import { createRoute } from '../../utils';
-import { DelayedLoadIndicator } from '../../custom-mui-components/delayed-load-indicator';
+import { DelayedLoadIndicator } from '../../components/delayed-load-indicator';
 import { ExtensionRegistryService } from '../../extension-registry-service';
 import { UserData } from '../../extension-registry-types';
 import { PageSettings } from '../../page-settings';
